backend/pyproject.toml
[project]
name = "mcp-use-visual"
version = "0.1.0"
description = "MCP-powered Ultimate AI Assistant - FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core web framework
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    # Environment variables
    "python-dotenv>=1.0.0",
    # LLM integration
    "langchain-openai>=0.3.28",
    "langchain-ollama>=0.3.5",
    # MCP protocol
    "mcp-use>=1.3.7",
    # Python 3.12 compatibility fix
    "orjson>=3.10.0",
]
